MAINT: clean up warnings in fiff

+++ src/libraries/fiff/fiff_dig_point_set.cpp
@@ -221,7 +221,7 @@ void FiffDigPointSet::writeToStream(FiffStream* p_pStream)
         p_pStream->write_dig_point(m_qListDigPoint[h]);
     }
 
-    printf("\t%d digitizer points written\n", m_qListDigPoint.size());
+    printf("\t%d digitizer points written\n", (int)m_qListDigPoint.size());
     p_pStream->end_block(FIFFB_ISOTRAK);
     p_pStream->end_block(FIFFB_MEAS_INFO);
     p_pStream->end_block(FIFFB_MEAS);
